+++ change_log.md
@@ -6,6 +6,7 @@
 * Split RPC methods to it's own namespace, currently include: `cfx`, `pos`, `trace`, `txpool`
 * Add support for batch RPC
 * Browser export class name change from Conflux to `TreeGraph`
+* Add method `getNextUsableNonce` to `conflux.advanced`, which will first try to use `txpool_nextNonce` to get a usable nonce, if failed it will fall back to `cfx_getNextNonce`
 
 ## v1.7.2